# apps/backend/swagger-output/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/ticket/org/{organisationId}": {
      "get": {
        "operationId": "TicketController_findAll",
        "parameters": [
          {
            "name": "organisationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "statusId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "agentId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "priority",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Ticket"
        ]
      }
    },
    "/ticket/org/{organisationId}/stats": {
      "get": {
        "operationId": "TicketController_getStats",
        "parameters": [
          {
            "name": "organisationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Ticket"
        ]
      }
    },
    "/ticket/{id}": {
      "get": {
        "operationId": "TicketController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Ticket"
        ]
      },
      "patch": {
        "operationId": "TicketController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Ticket"
        ]
      },
      "delete": {
        "operationId": "TicketController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Ticket"
        ]
      }
    },
    "/ticket": {
      "post": {
        "operationId": "TicketController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Ticket"
        ]
      }
    },
...
  },
  "info": {
    "title": "Bedones Agents API",
    "description": "API backend pour le CRM social Bedones",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
...
      "CreateTicketDto": {
        "type": "object",
        "properties": {
          "organisationId": {
            "type": "string"
          },
          "agentId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "statusId": {
            "type": "string"
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "URGENT"
            ]
          },
          "contactName": {
            "type": "string"
          },
          "contactId": {
            "type": "string"
          },
          "provider": {
            "type": "string",
            "enum": [
              "FACEBOOK",
              "INSTAGRAM",
              "WHATSAPP",
              "TIKTOK"
            ]
          },
          "conversationId": {
            "type": "string"
          },
          "assignedTo": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          }
        },
        "required": [
          "organisationId",
          "title"
        ]
      },
      "UpdateTicketDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "statusId": {
            "type": "string"
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "URGENT"
            ]
          },
          "assignedTo": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          }
        }
      },
...
    }
  }
}